# openapi.yaml
openapi: 3.0.0
info:
  title: Indicator Backend API
  version: 1.0.0
  description: |
    API documentation for the Indicator application backend.

    This API provides endpoints for managing educational institutions' data including:
    - User authentication and management
    - School basic information (alapadatok)
    - Student enrollment data (tanulo_letszam)
    - Educational data (tanugyi_adatok)
    - Competency measurements (kompetencia)
    - Admission statistics (felvettek_szama)
    - Table and data access management

    ## Authentication

    Most endpoints require JWT Bearer token authentication. To authenticate:
    1. Use the `/auth/login` endpoint to obtain a token
    2. Include the token in the Authorization header: `Bearer <token>`

    ## Base URL

    All API endpoints are prefixed with `/api/v1`

  contact:
    name: API Support
    url: https://indicator.pollak.info
    email: support@pollak.info
  license:
    name: Private
    url: https://indicator.pollak.info/license
servers:
  - url: /api/v1
    description: API v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for authentication
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error message
      example:
        error: "Invalid input data"
        message: "The provided data does not meet the required format"
...
    Dobbanto:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
          example: 1
        alapadatok_id:
          type: integer
          description: School identifier reference
          example: 1
        tanev_kezdete:
          type: integer
          description: School year start year
          example: 2023
        dobbanto_szama:
          type: integer
          description: Number of repeating students
          example: 15
        tanulok_osszesen:
          type: integer
          description: Total number of students
          example: 500
        datum:
          type: string
          format: date-time
          description: Date of record
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          nullable: true
          description: Last update timestamp
        alapadatok:
          type: object
          properties:
            nev:
              type: string
              description: School name
              example: "Pollák Antal Technikum"
    DobbantoInput:
      type: object
      required:
        - alapadatok_id
        - tanev_kezdete
        - dobbanto_szama
        - tanulok_osszesen
      properties:
        alapadatok_id:
          type: integer
          description: School identifier reference
          example: 1
        tanev_kezdete:
          type: integer
          description: School year start year
          example: 2023
        dobbanto_szama:
          type: integer
          description: Number of repeating students
          example: 15
        tanulok_osszesen:
          type: integer
          description: Total number of students
          example: 500

    # Elegedettseg schemas
...
security:
  - bearerAuth: []
...
paths:
...
  /dobbanto/{tanev}:
    get:
      tags: [Dobbanto]
      summary: Get dobbanto data by school year
      description: Retrieve dobbanto (repeating students) data for a specific school year and previous 4 years
      parameters:
        - in: path
          name: tanev
          required: true
          description: The school year end (e.g., 2024 for 2023-2024 school year)
          schema:
            type: integer
            example: 2024
      responses:
        "200":
          description: Dobbanto data retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dobbanto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /dobbanto/{alapadatokId}/{tanev}:
    get:
      tags: [Dobbanto]
      summary: Get dobbanto data by school and year
      description: Retrieve dobbanto data for a specific school and school year
      parameters:
        - in: path
          name: alapadatokId
          required: true
          description: The school identifier
          schema:
            type: integer
            example: 1
        - in: path
          name: tanev
          required: true
          description: The school year end
          schema:
            type: integer
            example: 2024
      responses:
        "200":
          description: Dobbanto data retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dobbanto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Dobbanto]
      summary: Delete dobbanto data by school and year
      description: Delete all dobbanto data for a specific school and school year
      parameters:
        - in: path
          name: alapadatokId
          required: true
          description: The school identifier
          schema:
            type: integer
            example: 1
        - in: path
          name: tanev
          required: true
          description: The school year end
          schema:
            type: integer
            example: 2024
      responses:
        "204":
          description: Dobbanto data deleted successfully
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /dobbanto:
    post:
      tags: [Dobbanto]
      summary: Create new dobbanto data
      description: Create new dobbanto (repeating students) data entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DobbantoInput"
      responses:
        "201":
          description: Dobbanto data created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dobbanto"
        "400":
          description: Bad request - Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /dobbanto/{id}:
    put:
      tags: [Dobbanto]
      summary: Update dobbanto data
      description: Update existing dobbanto (repeating students) data entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Dobbanto record ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DobbantoInput"
      responses:
        "200":
          description: Dobbanto data updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dobbanto"
        "400":
          description: Bad request - Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Record not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Elegedettseg endpoints
...